clothes_jxc/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
idna==2.8
Pillow==6.2.1
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
mysqlclient==2.0.1